requirements.txt
streamlit==1.50.0
openpyxl==3.1.5
google-genai==1.47.0
anthropic>=0.25.0
python-dotenv==1.2.1
